// app/view/models/issue.ts
export class Issue {
        amcatId: number;
        dateOfDrive : any;
        venueOfDrive : string;
        studentName : string;
        gender : string;
        yearOfGraduation: number;
        collegeName: string;
        university: string;
        studentCourse: string;
        studentBranch: string;
        markTenth : number;
        markTwelve : number;
        collegeMarks: number;
        collegePercentage : number;
        studentAddress : string;
        studentMobile: number;
        additionalContactNumber : number;
        studentEmail : string;
        addtionalEmailId : string;
        collegeAddress: string;
        nameOfPlacementOfficer : string;
        numberOfPlacementOfficer : number;
        emailOfPlacementOfficer: string;
        preLearningLink: string;
        litmusBatch : string;
        litmusStatus : string;
        studentRemarks: string;
        studentAdditionalRemarks: string;
        createdTime: any;
        modifiedTime: any;
        backgroundVerification : string;
        

}